+++ next.config.js
@@ -21,5 +21,11 @@ module.exports = withBundleAnalyzer(
       locales: ['en'],
       defaultLocale: 'en',
     },
+    rewrites: async () => [
+      {
+        source: '/sitemap.xml',
+        destination: '/api/sitemap.xml',
+      },
+    ],
   })
 );
